Relocate cache and remove core module
