--- src/bfscraper/core/__init__.py
@@ -1,5 +1,0 @@
-"""Core package.
-
-Author:
-    Paulo Sanchez (@erlete)
-"""
